+++ hapi-fhir-server/src/test/java/ca/uhn/fhir/rest/api/server/IBundleProviderTest.java
@@ -1,18 +1,11 @@
 package ca.uhn.fhir.rest.api.server;
 
 import ca.uhn.fhir.context.ConfigurationException;
-import ca.uhn.fhir.context.FhirVersionEnum;
 import ca.uhn.fhir.i18n.Msg;
+import ca.uhn.fhir.model.TestFhirResource;
 import ca.uhn.fhir.rest.server.SimpleBundleProvider;
 import com.google.common.collect.Lists;
-import org.hl7.fhir.instance.model.api.IAnyResource;
-import org.hl7.fhir.instance.model.api.IBaseMetaType;
-import org.hl7.fhir.instance.model.api.IBaseResource;
-import org.hl7.fhir.instance.model.api.IIdType;
-import org.hl7.fhir.instance.model.api.IPrimitiveType;
 import org.junit.jupiter.api.Test;
-
-import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -26,7 +19,7 @@ public class IBundleProviderTest {
 		SimpleBundleProvider provider = new SimpleBundleProvider();
 		assertTrue(provider.isEmpty());
 
-		provider = new SimpleBundleProvider(Lists.newArrayList(new TestResource()));
+		provider = new SimpleBundleProvider(Lists.newArrayList(new TestFhirResource()));
 		assertFalse(provider.isEmpty());
 	}
 
@@ -40,7 +33,7 @@ public class IBundleProviderTest {
 		};
 		assertTrue(provider.isEmpty());
 
-		provider = new SimpleBundleProvider(Lists.newArrayList(new TestResource())) {
+		provider = new SimpleBundleProvider(Lists.newArrayList(new TestFhirResource())) {
 			@Override
 			public Integer size() {
 				return null;
@@ -61,74 +54,6 @@ public class IBundleProviderTest {
 			provider.getAllResources();
 			fail();		} catch (ConfigurationException e) {
 			assertEquals(Msg.code(464) + "Attempt to request all resources from an asynchronous search result.  The SearchParameterMap for this search probably should have been synchronous.", e.getMessage());
-		}
-	}
-
-	private static class TestResource implements IAnyResource {
-
-		@Override
-		public boolean isEmpty() {
-			return false;
-		}
-
-		@Override
-		public boolean hasFormatComment() {
-			return false;
-		}
-
-		@Override
-		public List<String> getFormatCommentsPre() {
-			return null;
-		}
-
-		@Override
-		public List<String> getFormatCommentsPost() {
-			return null;
-		}
-
-		@Override
-		public Object getUserData(String theName) {
-			return null;
-		}
-
-		@Override
-		public IAnyResource setId(String theId) {
-			return null;
-		}
-
-		@Override
-		public void setUserData(String theName, Object theValue) {
-
-		}
-
-		@Override
-		public IBaseMetaType getMeta() {
-			return null;
-		}
-
-		@Override
-		public String getId() {
-			return null;
-		}
-
-		@Override
-		public IIdType getIdElement() {
-			return null;
-		}
-
-		@Override
-		public IPrimitiveType<String> getLanguageElement() {
-			return null;
-		}
-
-		@Override
-		public IBaseResource setId(IIdType theId) {
-			return null;
-		}
-
-		@Override
-		public FhirVersionEnum getStructureFhirVersionEnum() {
-			return null;
 		}
 	}
 }
